pyproject.toml
[project]
name = "ourgroceries-tool-py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "ourgroceries>=1.5.4",
    "python-dotenv>=1.1.1",
]
